fix(ollama): verify pulled model discovery during onboard
