+++ ci/test-file-size-budget.json
@@ -10,7 +10,7 @@
     "test/install-preflight.test.ts": 3934,
     "test/nemoclaw-start.test.ts": 4827,
     "test/onboard-messaging.test.ts": 2062,
-    "test/onboard-selection.test.ts": 6146,
+    "test/onboard-selection.test.ts": 6062,
     "test/onboard.test.ts": 4057,
     "test/policies.test.ts": 2332
   }
